+++ create-loop/a/src/aptitudes/index.ts
@@ -1,5 +1,0 @@
-import openHandler from './ui/openHandler';
-
-export {
-  openHandler,
-};