+++ create-loop/a/src/index.ts
@@ -1,8 +1,0 @@
-import {
-  openHandler,
-} from './aptitudes';
-
-(async function main(): Promise<void> {
-  console.log('Example Loop Started');
-  openHandler.start();
-})();