+++ create-loop/a/src/whispers/index.ts
@@ -1,4 +1,0 @@
-import IntroWhisper from './IntroWhisper';
-import IntroWhisperSimple from './IntroWhisperSimple';
-
-export { IntroWhisper, IntroWhisperSimple };