+++ loop-templates/templates/types.ts
@@ -11,6 +11,7 @@ export type TemplateFile = {
     | 'any'
     | 'nonzero'
     | 'clipboard'
+    | 'config'
     | 'document'
     | 'filesystem'
     | 'keyboard'