+++ vscode-extension/CHANGELOG.md
@@ -7,6 +7,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Adding Config aptitude
 ### Changed
 
 ## [0.4.0]